% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betweendaychanges.r
\name{covid_changes_between_days}
\alias{covid_changes_between_days}
\title{Covid Changes Between Days}
\usage{
covid_changes_between_days(day1, day2,region1)
}
\arguments{
\item{day1}{The first date of the bar chart that you wish to visualize and compare. Written in yyyy=mm-dd format.}

\item{day2}{The second date of the bar chart that you wish to visualize and compare. Written in yyyy=mm-dd format.}

\item{region1}{The region whose COVID-19 data you wish to visualize in the this comparative bar chart, written in string format.}

\item{region2}{to region 16 The other regions whose COVID-19 data you wish to visualize in the #'this comparative bar chart, written in string format.
Defaults to 'NULL'}
}
\value{
Two GG plot bar charts of the COVID-19 data, for days and regions selected
}
\description{
This function creates a bar chart using gg plot of the two dates selected, and #'plots the number of people who have died and been infected by COVID-19 in a specified region, #'on these two dates.
}
\note{
Regions must be written in full names
}
