% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_mapca.r
\name{covid_mapca}
\alias{covid_mapca}
\title{covid_mapca}
\usage{
covid_mapca(stats,token)
}
\arguments{
\item{token}{API token from api.pify
Defaults to "Tow8X4YNqnsWMFGbWxuPynzHh"}
}
\value{
return a Canada map with select statistics either infected count or deceased count
}
\description{
This function will return a Canada map with select statistics either infected count or deceased count
}
