% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_count.r
\name{daily_count}
\alias{daily_count}
\title{Daily Counts}
\usage{
daily_count()
}
\value{
a data frame with columns: 'region', 'infectedCount', 'deceasedCount', 'date', 'Daily_new_cases', and 'Daily_deaths', 'seven_day_average'.
}
\description{
This function creates a calculates the daily new cases and death counts from the pre-existing requesting data from the covid dataset.
Defaults to 'NULL'
}
