% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.r
\name{Get_data}
\alias{Get_data}
\title{Get_data}
\usage{
Get_data(apikey)
}
\arguments{
\item{apikey}{API token from Apify
Defaults to "Tow8X4YNqnsWMFGbWxuPynzHh"}
}
\value{
A data frame with 4 columns region,infectedCount,deceasedCount and the date
}
\description{
This function returns a well-formed data frame that contains basic Canadian COVID-19 information by region
}
