% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.r
\name{Getdata_syncing}
\alias{Getdata_syncing}
\title{Getdata_syncing}
\usage{
Getdata_syncing(apikey)
}
\arguments{
\item{apikey}{API token from apify
Defaults to "Tow8X4YNqnsWMFGbWxuPynzHh"}
}
\value{
A data frame with 4 columns region, infectedCount, deceasedCount and the date
}
\description{
This function will return a latest veresion of well-formed dataframe that contains basic Covid information
}
