% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseriesplot.r
\name{time_series_plot}
\alias{time_series_plot}
\title{Time Series Plot}
\usage{
time_series_plot(data, start_date, end_date,region1 )
}
\arguments{
\item{data}{character. The COVID-19 you wish to visualize. Use 'infected count', 'infected', #''deceased count' or
'deceased'.}

\item{start_date}{The start date of the time series you wish to visualize. Written in yyyy-mm-#'dd format.}

\item{end_date}{The end date of the time series you wish to visualize. Written in yyyy-mm-dd #'format.}

\item{region1}{The region you wish to visualize in the this time series plot, written in #'string format.}

\item{region2}{to region 16 The other regions you wish to visualize in the this time series #'plot written in string format.
Defaults to 'NULL'}
}
\value{
a GG plot time series of the date range, data and region selected
}
\description{
This function creates a time series plot using gg plot of the date range
selected, and plots the number of  people who have died or been infected by COVID-19 in a #'specified region.
}
\note{
Regions must be written in full names
}
