% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_mapca.r
\name{covid_mapca}
\alias{covid_mapca}
\title{covid_mapca}
\usage{
covid_mapca(stats,token)
}
\arguments{
\item{stats}{the data you want to visualize on the map, either 'infectedCount' or 'deceasedCount'}

\item{token}{API token from api.pify
Defaults to "Tow8X4YNqnsWMFGbWxuPynzHh"}
}
\value{
return a Canada map with select statistics either infected count or deceased count
}
\description{
This function will return a map of Canada with the selected data of either the current infected count or the current deceased count.
}
