% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currentdatabar.r
\name{current_data_barchart}
\alias{current_data_barchart}
\title{Current Data Bar CHart}
\usage{
current_data_barchart(token='4T9GEYHZ7PE9w8H29xynebW3L', data, region1, region2=NULL, region3=NULL,region4=NULL,region5=NULL,region6=NULL,region7=NULL,region8=NULL,region9=NULL,region10=NULL,region11=NULL,region12=NULL,region13=NULL,region14=NULL,region15=NULL,region16=NULL)
}
\arguments{
\item{token}{apikey used to access data'}

\item{data}{character. The COVID-19 you wish to visualize. Use 'infected count', 'infected', 'deceased count',
'deceased', 'both or 'Both'.}

\item{region1}{The region you wish to visualize in the this current data bar chart, written in string format.}

\item{region2}{The other regions you wish to visualize in this bar chart written in string format.
Defaults to 'NULL'}

\item{region3}{The other regions whose COVID-19 data you wish to visualize in the #'this comparative bar chart, written in string format.
Defaults to 'NULL'}

\item{region4}{The other regions whose COVID-19 data you wish to visualize in the #'this comparative bar chart, written in string format.
Defaults to 'NULL'}

\item{region5}{The other regions whose COVID-19 data you wish to visualize in the #'this comparative bar chart, written in string format.
Defaults to 'NULL'}

\item{region6}{The other regions whose COVID-19 data you wish to visualize in the #'this comparative bar chart, written in string format.
Defaults to 'NULL'}

\item{region7}{The other regions whose COVID-19 data you wish to visualize in the #'this comparative bar chart, written in string format.
Defaults to 'NULL'}

\item{region8}{The other regions whose COVID-19 data you wish to visualize in the #'this comparative bar chart, written in string format.
Defaults to 'NULL'}

\item{region9}{The other regions whose COVID-19 data you wish to visualize in the #'this comparative bar chart, written in string format.
Defaults to 'NULL'}

\item{region10}{The other regions whose COVID-19 data you wish to visualize in the #'this comparative bar chart, written in string format.
Defaults to 'NULL'}

\item{region11}{The other regions whose COVID-19 data you wish to visualize in the #'this comparative bar chart, written in string format.
Defaults to 'NULL'}

\item{region12}{The other regions whose COVID-19 data you wish to visualize in the #'this comparative bar chart, written in string format.
Defaults to 'NULL'}

\item{region13}{The other regions whose COVID-19 data you wish to visualize in the #'this comparative bar chart, written in string format.
Defaults to 'NULL'}

\item{region14}{The other regions whose COVID-19 data you wish to visualize in the #'this comparative bar chart, written in string format.
Defaults to 'NULL'}

\item{region15}{The other regions whose COVID-19 data you wish to visualize in the #'this comparative bar chart, written in string format.
Defaults to 'NULL'}

\item{region16}{The other regions whose COVID-19 data you wish to visualize in the #'this comparative bar chart, written in string format.
Defaults to 'NULL'}
}
\value{
a GG plot bar chart of the current COVID-19 data for the specified regions
}
\description{
This function creates a bar chart plot using gg plot, and plots a bar chart of the number of people who have died or been infected by COVID-19 to date in a specified region/regions.
}
\note{
Regions must be written in full names
}
