% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.r
\name{Get_data}
\alias{Get_data}
\title{Get_data}
\usage{
Get_data(apikey)
}
\arguments{
\item{apikey}{API token from api.pify
Defaults to "Tow8X4YNqnsWMFGbWxuPynzHh"}
}
\value{
A dataframe with 4 columns region,infectedCount,deceasedCount and date
}
\description{
This function will return a well-formed dataframe that contains basic Covid information
}
