% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.r
\name{Getdata_syncing}
\alias{Getdata_syncing}
\title{Getdata_syncin}
\usage{
Getdata_syncing(apikey)
}
\arguments{
\item{apikey}{API token from api.pify
Defaults to "Tow8X4YNqnsWMFGbWxuPynzHh"}
}
\value{
A dataframe with 4 columns region,infectedCount,deceasedCount and date
}
\description{
This function will return a latest veresion of well-formed dataframe that contains basic Covid information
}
