% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseriesplot.r
\name{time_series_plot}
\alias{time_series_plot}
\title{Time Series Plot}
\usage{
time_series_plot(apikey='4T9GEYHZ7PE9w8H29xynebW3L', data, start_date, end_date, region1, region2=NULL, region3=NULL,region4=NULL,region5=NULL,region6=NULL,region7=NULL,region8=NULL,region9=NULL,region10=NULL,region11=NULL,region12=NULL,region13=NULL,region14=NULL,region15=NULL,region16=NULL)
}
\arguments{
\item{apikey}{API token from api.pify}

\item{data}{character. The COVID-19 you wish to visualize. Use 'infected count', 'infected', 'deceased count' or
'deceased'.}

\item{start_date}{The start date of the time series you wish to visualize. Written in yyyy-mm-dd format.}

\item{end_date}{The end date of the time series you wish to visualize. Written in yyyy-mm-dd format.}

\item{region1}{The region you wish to visualize in the this time series plot, written in string format.}

\item{region2}{The other regions you wish to visualize in the this time series plot written in string format.
Defaults to 'NULL'}

\item{region3}{The other regions whose COVID-19 data you wish to visualize in the #'this comparative bar chart, written in string format.
Defaults to 'NULL'}

\item{region4}{The other regions whose COVID-19 data you wish to visualize in the #'this comparative bar chart, written in string format.
Defaults to 'NULL'}

\item{region5}{The other regions whose COVID-19 data you wish to visualize in the #'this comparative bar chart, written in string format.
Defaults to 'NULL'}

\item{region6}{The other regions whose COVID-19 data you wish to visualize in the #'this comparative bar chart, written in string format.
Defaults to 'NULL'}

\item{region7}{The other regions whose COVID-19 data you wish to visualize in the #'this comparative bar chart, written in string format.
Defaults to 'NULL'}

\item{region8}{The other regions whose COVID-19 data you wish to visualize in the #'this comparative bar chart, written in string format.
Defaults to 'NULL'}

\item{region9}{The other regions whose COVID-19 data you wish to visualize in the #'this comparative bar chart, written in string format.
Defaults to 'NULL'}

\item{region10}{The other regions whose COVID-19 data you wish to visualize in the #'this comparative bar chart, written in string format.
Defaults to 'NULL'}

\item{region11}{The other regions whose COVID-19 data you wish to visualize in the #'this comparative bar chart, written in string format.
Defaults to 'NULL'}

\item{region12}{The other regions whose COVID-19 data you wish to visualize in the #'this comparative bar chart, written in string format.
Defaults to 'NULL'}

\item{region13}{The other regions whose COVID-19 data you wish to visualize in the #'this comparative bar chart, written in string format.
Defaults to 'NULL'}

\item{region14}{The other regions whose COVID-19 data you wish to visualize in the #'this comparative bar chart, written in string format.
Defaults to 'NULL'}

\item{region15}{The other regions whose COVID-19 data you wish to visualize in the #'this comparative bar chart, written in string format.
Defaults to 'NULL'}

\item{region16}{The other regions whose COVID-19 data you wish to visualize in the #'this comparative bar chart, written in string format.
Defaults to 'NULL'}
}
\value{
a GG plot time series of the date range, data and regions selected
}
\description{
This function creates a time series plot using gg plot of the date range
selected, and plots the number of  people who have died or been infected by COVID-19 in a specified region/regions.
}
\note{
Regions must be written in full names
}
